Added style to the Home page
I worked in different images but I wasn't too happy. I left the dark
grey background while we find a proper image for the background.

You can try how a image looks like by uncommenting the background
property in the CSS, and changing the picture.

I have also change the #555 for #333 generally. As #555 was quite light.